P2/requirements.txt
Django==4.0.3
Pillow==9.0.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
django-rest-swagger==2.2.0
